requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.1
httpx>=0.25.2
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.4
plotly>=5.17.0
requests>=2.31.0
python-multipart>=0.0.6 